# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
/**
 * Writing style to apply when polishing
 */
export type AiPolishMode = (typeof AiPolishMode)[keyof typeof AiPolishMode];

export const AiPolishMode = {
  professional: "professional",
  technician: "technician",
  warranty: "warranty",
  "equipment-memory": "equipment-memory",
  "pm-summary": "pm-summary",
  "email-customer": "email-customer",
  "work-order": "work-order",
} as const;

export interface AiPolishBody {
  /** Raw field notes to polish */
  text: string;
  mode: AiPolishMode;
  /** Optional Clerk user ID for Pro status lookup */
  clientId?: string;
  /** When true, applies a speech-interpretation pass first to correct likely speech recognition errors (mishearings, phonetic substitutions) before applying the writing style. */
  fromVoice?: boolean;
}

export interface AiPolishResponse {
  /** Improved version of the input text */
  polished: string;
}

/**
 * A learned correction from a technician's personal vocabulary profile
 */
export interface UserVoiceCorrection {
  /** The word or phrase as the technician spoke it */
  original: string;
  /** The technician's preferred interpretation or spelling */
  preferred: string;
  /**
   * Number of times the technician has confirmed this preference
   * @minimum 1
   */
  count: number;
}

export interface VoiceInterpretBody {
  /** Raw speech-recognition transcript from the technician */
  rawTranscript: string;
  /** Optional personal vocabulary corrections learned from this technician's past usage */
  userCorrections?: UserVoiceCorrection[];
}

/**
 * A phrase the AI was less than 85% confident in interpreting
 */
export interface VoiceUncertainPhrase {
  /** The phrase exactly as it appeared in the raw transcript */
  original: string;
  /** The AI's HVAC interpretation of the phrase */
  suggested: string;
  /**
   * Confidence score for this specific phrase (0–100)
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
}

/**
 * Structured equipment memory data extracted from the interpreted transcript
 */
export interface VoiceMemoryExtracts {
  /** Parts or components that were replaced */
  componentsReplaced: string[];
  /** Repairs or corrective actions performed */
  repairsPerformed: string[];
  /** Refrigerant type if mentioned (e.g. R-22, R-410A) */
  refrigerantType?: string;
  /** Refrigerant amount with units if mentioned (e.g. 8 lb 12 oz) */
  refrigerantAmount?: string;
  /** Follow-up work recommended or required */
  followUp?: string;
  /** Preventive maintenance reminders extracted */
  pmReminders?: string;
  /** Notable conditions observed during the visit */
  observedConditions?: string;
  /** Any warranty-relevant information mentioned */
  warrantyInfo?: string;
}

/**
 * Full result from the HVAC Voice Intelligence pipeline
 */
export interface VoiceInterpretResult {
  /** The raw transcript exactly as received — never modified */
  original: string;
  /** Professional HVAC service documentation suitable for service records */
  professional: string;
  /** Plain-English summary suitable for building owners or facilities managers */
  customer: string;
  /**
   * Overall interpretation confidence score (0–100)
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
  /** Phrases the AI was less than 85% confident in interpreting */
  uncertainPhrases: VoiceUncertainPhrase[];
  memoryExtracts: VoiceMemoryExtracts;
}

export interface HealthStatus {
  status: string;
}

export interface DiagnoseHvacBody {
  /** Description of HVAC symptoms */
  symptoms: string;
  /** Anonymous client identifier for Pro status lookup */
  clientId?: string;
  /** Optional tester email for Apple/Google review bypass */
  testerEmail?: string;
}

/**
 * Urgency of repair
 */
export type HvacDiagnosisEntryPriorityLevel =
  (typeof HvacDiagnosisEntryPriorityLevel)[keyof typeof HvacDiagnosisEntryPriorityLevel];

export const HvacDiagnosisEntryPriorityLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

/**
 * A single matched diagnostic entry from the knowledge base
 */
export interface HvacDiagnosisEntry {
  /** Knowledge base entry ID */
  id: string;
  /** Specific diagnostic title (e.g. "Carrier RTU — Compressor Hard Start") */
  title: string;
  /** Broad fault category (e.g. "No Cool", "Trips Breaker") */
  category: string;
  /** Explanation of why this entry matched the symptoms */
  whyThisFits: string;
  /** Ranked list of likely root causes */
  likelyCauses: string[];
  /** Field-level checks a technician should perform first */
  firstChecks: string[];
  /** Specific meter and instrument measurements to take */
  meterChecks: string[];
  /** Urgency of repair */
  priorityLevel: HvacDiagnosisEntryPriorityLevel;
  /**
   * Match confidence percentage
   * @minimum 0
   * @maximum 100
   */
  confidencePercent: number;
  /** Recommended course of action */
  recommendedAction: string;
  /** Safety or operational risk if this issue is left unaddressed */
  riskNote: string;
}

/**
 * Top diagnosis match plus up to 2 alternatives
 */
export interface HvacDiagnosisResult {
  primary: HvacDiagnosisEntry;
  /** Up to 2 alternative diagnostic matches ranked by confidence */
  alternatives: HvacDiagnosisEntry[];
  /** Whether this response contains full Pro-tier data */
  isPro: boolean;
}