# lib/api-zod/src/generated/types/userVoiceCorrection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

/**
 * A learned correction from a technician's personal vocabulary profile
 */
export interface UserVoiceCorrection {
  /** The word or phrase as the technician spoke it */
  original: string;
  /** The technician's preferred interpretation or spelling */
  preferred: string;
  /**
   * Number of times the technician has confirmed this preference
   * @minimum 1
   */
  count: number;
}