# lib/api-zod/src/generated/types/voiceInterpretBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { UserVoiceCorrection } from "./userVoiceCorrection";

export interface VoiceInterpretBody {
  /** Raw speech-recognition transcript from the technician */
  rawTranscript: string;
  /** Optional personal vocabulary corrections learned from this technician's past usage */
  userCorrections?: UserVoiceCorrection[];
}