# lib/api-zod/src/generated/types/voiceInterpretResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { VoiceMemoryExtracts } from "./voiceMemoryExtracts";
import type { VoiceUncertainPhrase } from "./voiceUncertainPhrase";

/**
 * Full result from the HVAC Voice Intelligence pipeline
 */
export interface VoiceInterpretResult {
  /** The raw transcript exactly as received — never modified */
  original: string;
  /** Professional HVAC service documentation suitable for service records */
  professional: string;
  /** Plain-English summary suitable for building owners or facilities managers */
  customer: string;
  /**
   * Overall interpretation confidence score (0–100)
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
  /** Phrases the AI was less than 85% confident in interpreting */
  uncertainPhrases: VoiceUncertainPhrase[];
  memoryExtracts: VoiceMemoryExtracts;
}