# lib/api-zod/src/generated/types/voiceMemoryExtracts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

/**
 * Structured equipment memory data extracted from the interpreted transcript
 */
export interface VoiceMemoryExtracts {
  /** Parts or components that were replaced */
  componentsReplaced: string[];
  /** Repairs or corrective actions performed */
  repairsPerformed: string[];
  /** Refrigerant type if mentioned (e.g. R-22, R-410A) */
  refrigerantType?: string;
  /** Refrigerant amount with units if mentioned (e.g. 8 lb 12 oz) */
  refrigerantAmount?: string;
  /** Follow-up work recommended or required */
  followUp?: string;
  /** Preventive maintenance reminders extracted */
  pmReminders?: string;
  /** Notable conditions observed during the visit */
  observedConditions?: string;
  /** Any warranty-relevant information mentioned */
  warrantyInfo?: string;
}