# lib/api-zod/src/generated/types/voiceUncertainPhrase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

/**
 * A phrase the AI was less than 85% confident in interpreting
 */
export interface VoiceUncertainPhrase {
  /** The phrase exactly as it appeared in the raw transcript */
  original: string;
  /** The AI's HVAC interpretation of the phrase */
  suggested: string;
  /**
   * Confidence score for this specific phrase (0–100)
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
}